Fix form validation for Ambulatory Sites
The Add/Edit forms will no longer close with no submission when they
contain invalid data. This was caused by submit being wrapped in a
DialogClose, now DialogClose is manually invoked when the form passes
validation with checkValidation. To accomplish this, the location text
fields need to be not readOnly but with a change that does nothing, so
they are essentially still readOnly but they have validation run on them